made some changes, cleaned up the documents and fixed the installation script.
NtCreateFile still not working, however instead of direct crash, the files are create .. like the donut has log.txt and then 2 not valid signs.. 
Probably there is something wrong with unicode.

when drag drop xbe files on dxbx.exe for the first time, the default dll was still cxbx. now when drag drop xbe files it is always converted to exe with the dxbxkrnl.dll