Updated DirectX wrappers (we where using an older version apparently).
PS: Some of these units contain the "inline" directive. But because we don't use these units, this should not harm our support for various Delphi compiler versions.
As a matter of fact, we're probably using this directive ourselves too - we should probably make these conditional to restore Delphi 7 compatibility. *shrug*